refactor(cmd/pro): split functions across file to reduce file size
